requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
datasets==2.21.0
Deprecated==1.2.14
dill==0.3.8
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
multidict==6.0.5
multiprocess==0.70.16
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.7
